VEC-577: list the four new current tables in the schema register so the schema check passes
The schema conformance check compares every live table against the column
register and failed because the four current tables were missing from it: a
table that exists in the database but not in the register is reported as
unregistered.

They are listed now, as a new "derived" kind: rows computed from another
governed table rather than read from the chain. Their columns still have to
match the shared column vocabulary, but the rules that only make sense for
freshly ingested data (needing a chain, a protocol and an observation time, and
being carried into the transformed layer) do not apply to them.

Two columns on token_price_current were also widened to the standard widths for
an oracle reference and a block version. They had inherited the narrower ones
from the price history, which predates that convention, and the register wants
the standard widths on a table created after it. Everything the column is
compared against elsewhere already uses the wider type, so this removes a
mismatch rather than adding one.

+++ stl-verify/db/migrations/20260820_120000_create_current_position_tables.sql
@@ -214,13 +214,17 @@ ORDER BY srd.protocol_id, srd.token_id,
 -- token_price_current
 -- ============================================================================
 
+-- oracle_id and block_version are the canonical widths (int8 / int4) here, not the
+-- narrower ones onchain_token_price kept from before the convention: the widths this
+-- table joins against (oracle.id, protocol_oracle.oracle_id) are BIGINT, and the
+-- schema_master register requires canonical widths on a table created after it.
 CREATE TABLE IF NOT EXISTS token_price_current (
     protocol_id        BIGINT         NOT NULL,
     token_id           BIGINT         NOT NULL,
     price_usd          NUMERIC(30,18) NOT NULL,
-    oracle_id          SMALLINT       NOT NULL,
+    oracle_id          BIGINT         NOT NULL,
     block_number       BIGINT         NOT NULL,
-    block_version      SMALLINT       NOT NULL,
+    block_version      INT            NOT NULL,
     processing_version INT            NOT NULL,
     PRIMARY KEY (protocol_id, token_id)
 );
